requirements.txt
uvicorn[standard]==0.38.0
fastapi==0.121.2
python-dotenv==1.2.1

langgraph==1.0.3
langchain==1.0.7
langchain-openai==1.0.3

sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.1

pydantic==2.12.4
pydantic[email]==2.12.4